+++ CHANGELOG.md
@@ -7,7 +7,7 @@
 * Feature: [group] with type "htgroup", "none", "from_auth" (NEW)
 * Extension: [auth] type "pam": set groups of user to be used later
 * Adjustment: reject usernames starting or ending with "@" or having more than one "@"
-* Adjustment: reject usernames containing ":"
+* Adjustment: reject usernames containing ":" or ","
 * Add: [group] group_collections_folder replacing hardcoded "GROUPS"
 * Fix: pathutils: detect missing permission to create symbolic link on Microsoft Windows